LPS-105341 Moving to layout-seo-web
