+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/constants/LayoutAdminWebKeys.java
@@ -29,7 +29,4 @@ public class LayoutAdminWebKeys {
 	public static final String LAYOUT_PAGE_LAYOUT_ADMIN_DISPLAY_CONTEXT =
 		"LAYOUT_PAGE_LAYOUT_ADMIN_DISPLAY_CONTEXT";
 
-	public static final String LAYOUT_PAGE_LAYOUT_SEO_DISPLAY_CONTEXT =
-		"LAYOUT_PAGE_LAYOUT_SEO_DISPLAY_CONTEXT";
-
 }
