+++ modules/apps/layout/layout-seo-web/src/main/java/com/liferay/layout/seo/web/internal/display/context/LayoutsSEODisplayContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.admin.web.internal.display.context;
+package com.liferay.layout.seo.web.internal.display.context;
 
 import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.document.library.util.DLURLHelper;
