+++ modules/apps/layout/layout-seo-web/src/main/java/com/liferay/layout/seo/web/internal/portlet/action/EditOpenGraphMVCActionCommand.java
@@ -12,31 +12,26 @@
  * details.
  */
 
-package com.liferay.layout.admin.web.internal.portlet.action;
+package com.liferay.layout.seo.web.internal.portlet.action;
 
 import com.liferay.layout.admin.constants.LayoutAdminPortletKeys;
+import com.liferay.layout.seo.model.LayoutSEOEntry;
 import com.liferay.layout.seo.service.LayoutSEOEntryLocalService;
 import com.liferay.layout.seo.service.LayoutSEOEntryService;
 import com.liferay.portal.events.EventsProcessorUtil;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutConstants;
 import com.liferay.portal.kernel.model.LayoutTypePortlet;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.service.LayoutLocalService;
-import com.liferay.portal.kernel.service.LayoutService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextFactory;
 import com.liferay.portal.kernel.servlet.MultiSessionMessages;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LocalizationUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.PropertiesParamUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 
@@ -56,35 +51,21 @@ import org.osgi.service.component.annotations.Reference;
 	immediate = true,
 	property = {
 		"javax.portlet.name=" + LayoutAdminPortletKeys.GROUP_PAGES,
-		"mvc.command.name=/layout/edit_seo"
+		"mvc.command.name=/layout/edit_open_graph"
 	},
 	service = MVCActionCommand.class
 )
-public class EditSEOMVCActionCommand extends BaseMVCActionCommand {
+public class EditOpenGraphMVCActionCommand extends BaseMVCActionCommand {
 
 	@Override
 	protected void doProcessAction(
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
-		UploadPortletRequest uploadPortletRequest =
-			_portal.getUploadPortletRequest(actionRequest);
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
 		long groupId = ParamUtil.getLong(actionRequest, "groupId");
 		boolean privateLayout = ParamUtil.getBoolean(
 			actionRequest, "privateLayout");
 		long layoutId = ParamUtil.getLong(actionRequest, "layoutId");
-		Map<Locale, String> titleMap = LocalizationUtil.getLocalizationMap(
-			actionRequest, "title");
-		Map<Locale, String> descriptionMap =
-			LocalizationUtil.getLocalizationMap(actionRequest, "description");
-		Map<Locale, String> keywordsMap = LocalizationUtil.getLocalizationMap(
-			actionRequest, "keywords");
-		Map<Locale, String> robotsMap = LocalizationUtil.getLocalizationMap(
-			actionRequest, "robots");
 
 		ServiceContext serviceContext = ServiceContextFactory.getInstance(
 			Layout.class.getName(), actionRequest);
@@ -92,89 +73,74 @@ public class EditSEOMVCActionCommand extends BaseMVCActionCommand {
 		Layout layout = _layoutLocalService.getLayout(
 			groupId, privateLayout, layoutId);
 
-		layout = _layoutService.updateLayout(
-			groupId, privateLayout, layoutId, layout.getParentLayoutId(),
-			layout.getNameMap(), titleMap, descriptionMap, keywordsMap,
-			robotsMap, layout.getType(), layout.isHidden(),
-			layout.getFriendlyURLMap(), layout.isIconImage(), null,
-			serviceContext);
+		boolean openGraphDescriptionEnabled = ParamUtil.getBoolean(
+			actionRequest, "openGraphDescriptionEnabled");
+		Map<Locale, String> openGraphDescriptionMap =
+			LocalizationUtil.getLocalizationMap(
+				actionRequest, "openGraphDescription");
+		Map<Locale, String> openGraphImageAltMap =
+			LocalizationUtil.getLocalizationMap(
+				actionRequest, "openGraphImageAlt");
+		long openGraphImageFileEntryId = ParamUtil.getLong(
+			actionRequest, "openGraphImageFileEntryId");
+		boolean openGraphTitleEnabled = ParamUtil.getBoolean(
+			actionRequest, "openGraphTitleEnabled");
+		Map<Locale, String> openGraphTitleMap =
+			LocalizationUtil.getLocalizationMap(
+				actionRequest, "openGraphTitle");
 
-		boolean canonicalURLEnabled = ParamUtil.getBoolean(
-			actionRequest, "canonicalURLEnabled");
-		Map<Locale, String> canonicalURLMap =
-			LocalizationUtil.getLocalizationMap(actionRequest, "canonicalURL");
+		LayoutSEOEntry layoutSEOEntry =
+			_layoutSEOEntryLocalService.fetchLayoutSEOEntry(
+				groupId, privateLayout, layoutId);
+
+		boolean canonicalURLEnabled = false;
+
+		if ((layoutSEOEntry != null) &&
+			layoutSEOEntry.isCanonicalURLEnabled()) {
+
+			canonicalURLEnabled = true;
+		}
+
+		Map<Locale, String> canonicalURLMap = null;
+
+		if (layoutSEOEntry != null) {
+			canonicalURLMap = layoutSEOEntry.getCanonicalURLMap();
+		}
 
 		_layoutSEOEntryService.updateLayoutSEOEntry(
 			groupId, privateLayout, layoutId, canonicalURLEnabled,
-			canonicalURLMap, serviceContext);
+			canonicalURLMap, openGraphDescriptionEnabled,
+			openGraphDescriptionMap, openGraphImageAltMap,
+			openGraphImageFileEntryId, openGraphTitleEnabled, openGraphTitleMap,
+			serviceContext);
 
 		Layout draftLayout = _layoutLocalService.fetchLayout(
 			_portal.getClassNameId(Layout.class), layout.getPlid());
 
 		if (draftLayout != null) {
-			_layoutService.updateLayout(
-				groupId, privateLayout, draftLayout.getLayoutId(),
-				draftLayout.getParentLayoutId(), draftLayout.getNameMap(),
-				titleMap, descriptionMap, keywordsMap, robotsMap,
-				draftLayout.getType(), draftLayout.isHidden(),
-				draftLayout.getFriendlyURLMap(), draftLayout.isIconImage(),
-				null, serviceContext);
-
 			_layoutSEOEntryService.updateLayoutSEOEntry(
 				groupId, privateLayout, draftLayout.getLayoutId(),
-				canonicalURLEnabled, canonicalURLMap, serviceContext);
+				canonicalURLEnabled, canonicalURLMap,
+				openGraphDescriptionEnabled, openGraphDescriptionMap,
+				openGraphImageAltMap, openGraphImageFileEntryId,
+				openGraphTitleEnabled, openGraphTitleMap, serviceContext);
 		}
-
-		themeDisplay.clearLayoutFriendlyURL(layout);
-
-		UnicodeProperties layoutTypeSettingsProperties =
-			layout.getTypeSettingsProperties();
-
-		UnicodeProperties formTypeSettingsProperties =
-			PropertiesParamUtil.getProperties(
-				actionRequest, "TypeSettingsProperties--");
 
 		LayoutTypePortlet layoutTypePortlet =
 			(LayoutTypePortlet)layout.getLayoutType();
 
-		String type = layout.getType();
-
-		if (type.equals(LayoutConstants.TYPE_PORTLET)) {
-			layoutTypeSettingsProperties.putAll(formTypeSettingsProperties);
-
-			boolean layoutCustomizable = GetterUtil.getBoolean(
-				layoutTypeSettingsProperties.get(
-					LayoutConstants.CUSTOMIZABLE_LAYOUT));
-
-			if (!layoutCustomizable) {
-				layoutTypePortlet.removeCustomization(
-					layoutTypeSettingsProperties);
-			}
-
-			layout = _layoutService.updateLayout(
-				groupId, privateLayout, layoutId,
-				layoutTypeSettingsProperties.toString());
-		}
-		else {
-			layoutTypeSettingsProperties.putAll(formTypeSettingsProperties);
-
-			layoutTypeSettingsProperties.putAll(
-				layout.getTypeSettingsProperties());
-
-			layout = _layoutService.updateLayout(
-				groupId, privateLayout, layoutId,
-				layoutTypeSettingsProperties.toString());
-		}
-
 		EventsProcessorUtil.process(
 			PropsKeys.LAYOUT_CONFIGURATION_ACTION_UPDATE,
 			layoutTypePortlet.getConfigurationActionUpdate(),
-			uploadPortletRequest,
+			_portal.getHttpServletRequest(actionRequest),
 			_portal.getHttpServletResponse(actionResponse));
 
 		String redirect = ParamUtil.getString(actionRequest, "redirect");
 
 		if (Validator.isNull(redirect)) {
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)actionRequest.getAttribute(WebKeys.THEME_DISPLAY);
+
 			redirect = _portal.getLayoutFullURL(layout, themeDisplay);
 		}
 
@@ -188,9 +154,6 @@ public class EditSEOMVCActionCommand extends BaseMVCActionCommand {
 	}
 
 	@Reference
-	private ActionUtil _actionUtil;
-
-	@Reference
 	private LayoutLocalService _layoutLocalService;
 
 	@Reference
@@ -198,9 +161,6 @@ public class EditSEOMVCActionCommand extends BaseMVCActionCommand {
 
 	@Reference
 	private LayoutSEOEntryService _layoutSEOEntryService;
-
-	@Reference
-	private LayoutService _layoutService;
 
 	@Reference
 	private Portal _portal;
