+++ modules/apps/layout/layout-seo-web/src/main/resources/META-INF/resources/init.jsp
@@ -20,15 +20,34 @@
 
 <%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
 taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
+taglib uri="http://liferay.com/tld/react" prefix="react" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.layout.seo.model.LayoutSEOSite" %><%@
+<%@ page import="com.liferay.layout.admin.kernel.model.LayoutTypePortletConstants" %><%@
+page import="com.liferay.layout.seo.model.LayoutSEOEntry" %><%@
+page import="com.liferay.layout.seo.model.LayoutSEOSite" %><%@
 page import="com.liferay.layout.seo.web.internal.display.context.OpenGraphSettingsDisplayContext" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
+page import="com.liferay.portal.kernel.exception.SitemapChangeFrequencyException" %><%@
+page import="com.liferay.portal.kernel.exception.SitemapIncludeException" %><%@
+page import="com.liferay.portal.kernel.exception.SitemapPagePriorityException" %><%@
+page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
+page import="com.liferay.portal.kernel.model.Layout" %><%@
+page import="com.liferay.portal.kernel.model.LayoutConstants" %><%@
+page import="com.liferay.portal.kernel.util.GetterUtil" %><%@
+page import="com.liferay.portal.kernel.util.HashMapBuilder" %><%@
+page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
 page import="com.liferay.portal.kernel.util.PortalUtil" %><%@
-page import="com.liferay.portal.kernel.util.Validator" %>
+page import="com.liferay.portal.kernel.util.StringUtil" %><%@
+page import="com.liferay.portal.kernel.util.UnicodeProperties" %><%@
+page import="com.liferay.portal.kernel.util.Validator" %><%@
+page import="com.liferay.portal.kernel.util.WebKeys" %><%@
+page import="com.liferay.portal.util.PropsValues" %>
+
+<%@ page import="java.util.Collections" %><%@
+page import="java.util.Map" %>
 
 <liferay-frontend:defineObjects />
