+++ modules/apps/layout/layout-seo-web/src/main/resources/META-INF/resources/js/seo/PreviewSeoEvents.es.js
